Support transformed coordinates
This adds a new handler, `TransformedSphericalCoordinateHandler`,
for coordinate systems with some transformation function between
their native coordinates (e.g. X1, X2, X3) and desired output
coordinate system (e.g. r, th, phi).  This is common in GRMHD/
black hole accretion codes, and is a prerequisite to using yt
with those codes (the preliminary support targets are the
Parthenon-based codes KHARMA and Phoebus).

In order to test the new system, it also adds support for parsing
KHARMA output via the porposed Parthenon frontend, including
determining and assigning an appropriate `CoordinateTransform`
object to the `.transform` dataset member.  The transforms
implemented in KHARMA are included as a part of the Parthenon
frontend.

Details are still very rough: in particular, vectors are not yet
supported at all and sightlines/1D integration support is
almost certainly broken.  And of course, anything not working
in spherical coordinates also doesn't work in the transformed
version.

Note an enumerated value for a `TransformedCartesian` coordinate
system is added, but this will error as such a system is not
implemented.  It will be useful in the future for e.g. snake
coordinates used to test GRMHD codes, but no testing files are
available from KHARMA to implement it.  Implementation should
be straightforward, using the same set of changes as in the
`TransformedSpherical` system vs `Spherical`